Updated code
Had to disable some of Chris' state switching override code and go back to the manual version, because of RF and the ACK packets.  Basically the radio buffer would get confused if we weren't checking for an ACK.  I should probably re-write the getStatus thing to not have the publicly accessible statbuf...it's horrible.

Functionality with the actual button, rotary encoder, etc. was a little finicky, and occasionally the board seems to lock up when switching states (screen code?).  It might also just be the wiring mess and such, which the PCB will hopefully fix.

I think we may be missing pinMode() calls in setup().

So far the rocket can (if ground test override applied) make it to flight, transmit data, and then transmit a status message when it is in recovery/landed.  To do next: add GPS position transmission from rocket when in recovery, and receive this on the ground side.